requirements.txt

numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
joblib==1.3.2
matplotlib==3.8.3
PyYAML==6.0.1
pytest==8.0.2
Flask==3.0.2
Werkzeug==3.0.1
seaborn==0.13.2
